+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - refactor CLI projects in MVC
 
 ## 2019-10-07
+### Added
+- testing for spring-cli-jdbc controller
+
 ### Fixed
 - regression in spring-cli-jdbc due to refactoring